requirements.txt
python-docx==1.1.0
jinja2==3.1.2
PyYAML==6.0.1
